requirements.txt
Flask==2.2.2
werkzeug==2.2.2
python-dotenv==0.21.0
gunicorn==20.1.0
openai==0.27.7
